Add compatibility for Rails 5
Moved prism-rails.js to prism.js and removed require at bottom. This is
because loading the same name as the gem seems to load all files
regardless of necessity. Additionally, prism.js was overwriting the
languages already loaded and isn't needed as it's the same as
prism/prism-core.js.

Finally, added a test for Turbolinks (5) so that the highlighter
continues to work across page reloads.